consolidate require's into a single loader file for the gem
In trying to setup component previews in the
actionview-component-demo repo, I realized that we were
requiring users of this gem to explicitly require the railtie
in order for them to work.

This PR moves us towards a more typical gem structure
where a single entry file, lib/action_view/component.rb,
loads the rest of the gem.
