+++ lib/action_view/component.rb
@@ -1,0 +1,3 @@
+require "action_view/component/monkey_patch"
+require "action_view/component/base"
+require "action_view/component/railtie"
