+++ lib/action_view/component/base.rb
@@ -1,38 +1,13 @@
 # frozen_string_literal: true
 
 require "active_model"
+require "action_view"
 require "active_support/configurable"
 require_relative "preview"
 
-# Monkey patch ActionView::Base#render to support ActionView::Component
-#
-# A version of this monkey patch was upstreamed in https://github.com/rails/rails/pull/36388
-# We'll need to upstream an updated version of this eventually.
-class ActionView::Base
-  module RenderMonkeyPatch
-    def render(options = {}, args = {}, &block)
-      if options.respond_to?(:render_in)
-        ActiveSupport::Deprecation.warn(
-          "passing component instances (`render MyComponent.new(foo: :bar)`) has been deprecated and will be removed in v2.0.0. Use `render MyComponent, foo: :bar` instead."
-        )
-
-        options.render_in(self, &block)
-      elsif options.is_a?(Class) && options < ActionView::Component::Base
-        options.new(args).render_in(self, &block)
-      elsif options.is_a?(Hash) && options.has_key?(:component)
-        options[:component].new(options[:locals]).render_in(self, &block)
-      else
-        super
-      end
-    end
-  end
-
-  prepend RenderMonkeyPatch
-end
-
 module ActionView
   module Component
-    class Base < ActionView::Base
+    class Base < ::ActionView::Base
       include ActiveModel::Validations
       include ActiveSupport::Configurable
       include ActionView::Component::Previews
